imorove filtering , after clicking search button , every filter props is going to complex.jsx
